+++ NEWS.md
@@ -1,5 +1,13 @@
 # lazyraster 0.1.0
 
+## BREAKING CHANGES
+
+* lazycrop is now removed, we can use `crop()` directly. 
+
+## CHANGES
+
+* S4 method `crop()` is now available for lazyraster. 
+
 * Function `as_raster()` gains a `native` argument to override dimension. 
 
 * Depend on vapour 0.4.0, because change in list return from read raster.